Move flask server code to root directory.
