+++ .gitignore
@@ -19,4 +19,4 @@ venv/
 __pycache__
 work
 
-server/output/
+output/
