+++ lib/spec_utils.py
