+++ tools/audio_remove.py
@@ -4,8 +4,8 @@ import soundfile as sf
 import torch
 from tqdm import tqdm
 
-from server.lib import dataset
-from server.lib import nets, spec_utils
+from lib import dataset
+from lib import nets, spec_utils
 
 AUDIO_REMOVE_DEVICE = "gpu"
 AUDIO_REMOVE_FFT_SIZE = 2048
